Fix: release script should use `make image-local`
`make build-local` doesn't build create a container image, it simply
builds the binaries. `make image-local` takes those binaries and creates
an image with them.

The release script needs git-chglog to create the CHANGELOG.md file, and
instead of assuming it's available locally, it uses the one that is part
of the container image. It's a bit heavy-handed, but it gets the job
done.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>