fix: Tweak golangci-lint wrapper
Erroring out if there's no configuration file does not match
golangci-lint's behavior. If it doesn't have a configuration file, it
uses a built-in default configuration.

So if there's no configuration file, just use v2, because there's no
need for a migration. This also gives a smoother transition path: if you
have a v1 configuration file, migrate it and this wrapper will
automatically switch to v2. If there's no configuration file, it uses
v2, too...

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>